MEGA65_R2: Removed everything
After starting the project we realized, that the R2's FPGA is too weak to host the M2M framework PLUS the actual core. We tried the C64 and it became a mess. We decided the the low amount of R2's ever manufactured (around 20) in conjunction with the efforts necessary to make it happen are not in balance. Also "make it happen" would have meant: Way less C64 features (and M2M features) for the R2, lower visual quality due to "no ascal" and other things. See also https://github.com/sy2002/MiSTer2MEGA65/issues/27