M2M: Fix regression bug (HDMI not showing)
The fundamental reason is that with the new
dynamic HDMI clock frequency selection (in
M2M/vhdl/controllers/HDMI/video_out_clock.vhd), the MMCM controlling the
HDMI clock will be temporarily brought into reset, while the frequency
is changing. However, we don't want the Core to reset, just because the
user is changing the screen
resolution. Therefore, we can not reset the ascal'er or the HyperRAM
either, because these three must all be reset simultaneously. This
latter fact is related to the interface to the physical HyperRAM device
is stateful, as well as the Avalon Memory Map interface between the
ascal'er and the HyperRAM controller.

So in all, the ascal'er, the Core, and the HyperRAM controller must all
just ignore any resets occurring with the HDMI clock. This is perfectly
safe, given that the impact seen by, e.g., the ascal'er is just that
there is a period in time, when the HDMI clock is not ticking.